Added method MiniMax
