+++ src/tictactoe/Application.java
@@ -9,9 +9,9 @@ import java.util.Scanner;
 import java.util.regex.Pattern;
 
 public final class Application {
-    private static final Pattern PATTERN_COMMAND = Pattern.compile("(exit|start( easy| medium| hard| mm| user){2})");
+    private static final Pattern PATTERN_COMMAND = Pattern.compile("(exit|start( easy| medium| hard| user){2})");
     private static final Map<String, Player> players = Map.of("user", User::new,
-            "easy", Easy::new, "medium", Medium::new, "hard", Hard::new, "mm", MiniMax::new);
+            "easy", Easy::new, "medium", Medium::new, "hard", Hard::new);
 
     void run() {
         do {
