+++ src/tictactoe/ai/Easy.java
@@ -3,16 +3,13 @@ package tictactoe.ai;
 import tictactoe.engine.Mark;
 import tictactoe.engine.Board;
 
-import java.util.Random;
 import java.util.logging.Logger;
 
 public class Easy extends Ai {
-    private final Random random;
     private static final Logger log = Logger.getLogger(Easy.class.getName());
 
     public Easy(Board board, Mark mark) {
         super(board, mark);
-        this.random = new Random();
         log.info("Easy Ai created");
     }
